fix(config): fix disk cache int overflow, vpn killswitch toggle, and snake_case env vars (fixes #597)

+++ src/Leecharr.Api.V1/Config/BitTorrentConfigResource.cs
@@ -89,7 +89,7 @@ public class BitTorrentConfigResource : RestResource
     public int PeerPortRandomHigh { get; set; }
 
     // MonoTorrent Specific
-    public int DiskCacheBytes { get; set; }
+    public long DiskCacheBytes { get; set; }
 
     public string DiskCachePolicy { get; set; }
